requirements.txt
playwright==1.35.0
pillow==9.5.0
flask==2.2.5 